GCP/requirements.txt
apache-beam==2.14.0
avro==1.9.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
crcmod==1.7
dill==0.2.9
docopt==0.6.2
enum34==1.1.6
fastavro==0.21.24
fasteners==0.15
funcsigs==1.0.2
future==0.17.1
futures==3.3.0
google-api-core==1.14.2
google-apitools==0.5.28
google-auth==1.6.3
google-cloud-bigquery==1.6.1
google-cloud-bigtable==0.32.2
google-cloud-core==0.29.1
google-cloud-datastore==1.7.4
google-cloud-pubsub==0.39.1
google-cloud-storage==1.17.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
googledatastore==7.0.2
grpc-google-iam-v1==0.11.4
grpcio==1.22.0
hdfs==2.5.8
httplib2==0.12.0
idna==2.8
mock==2.0.0
monotonic==1.5
numpy==1.16.4
oauth2client==3.0.0
pandas==0.24.2
pbr==5.4.1
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.9.0
pyarrow==0.14.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pydot==1.4.1
pymongo==3.8.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyVCF==0.6.8
PyYAML==3.13
requests==2.22.0
rsa==4.0
scikit-learn==0.20.4
scipy==1.2.2
six==1.12.0
sklearn==0.0
typing==3.6.6
urllib3==1.25.3
